src/test/java/me/kktrkkt/studyolle/study/StudyControllerTest.java
package me.kktrkkt.studyolle.study;

import me.kktrkkt.studyolle.account.AccountRepository;
import me.kktrkkt.studyolle.account.WithAccount;
import me.kktrkkt.studyolle.infra.MockMvcTest;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;

import java.util.Optional;

import static org.junit.jupiter.api.Assertions.*;
import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@MockMvcTest
class StudyControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @Autowired
    private AccountRepository accounts;

    @Autowired
    private StudyRepository studys;

    @AfterEach
    void afterEach() {
        accounts.deleteAll();
    }

    @DisplayName("스터디 생성 폼 조회")
    @Test
    @WithAccount("user1")
    void StudyCreateForm() throws Exception {
        this.mockMvc.perform(get(StudyController.NEW_STUDY_URL))
                .andExpect(status().isOk())
                .andExpect(model().attributeExists("studySubmitForm"))
                .andExpect(view().name(StudyController.NEW_STUDY_VIEW))
                .andDo(print());
    }

    @DisplayName("스터디 생성 - 성공")
    @Test
    @WithAccount("user1")
    void createStudy_success() throws Exception {
        String url = "new-study";
        String title = "new-study";
        String bio = "bio";
        String explanation = "explanation";
        this.mockMvc.perform(post(StudyController.NEW_STUDY_URL)
                        .with(csrf())
                        .param("url", url)
                        .param("title", title)
                        .param("bio", bio)
                        .param("explanation", explanation)
                        .contentType(MediaType.APPLICATION_FORM_URLENCODED))
                .andExpect(status().is3xxRedirection())
                .andExpect(flash().attributeExists("study"))
                .andDo(print());
        Optional<Study> byUrl = studys.findByUrl(url);
        assertTrue(byUrl.isPresent());
    }

    @DisplayName("스터디 생성 - 실패")
    @Test
    @WithAccount("user1")
    void createStudy_failure() throws Exception {
        String url = "new/study";
        String title = "new-study";
        String bio = "bio";
        String explanation = "explanation";
        this.mockMvc.perform(post(StudyController.NEW_STUDY_URL)
                        .with(csrf())
                        .param("url", url)
                        .param("title", title)
                        .param("bio", bio)
                        .param("explanation", explanation)
                        .contentType(MediaType.APPLICATION_FORM_URLENCODED))
                .andExpect(status().isOk())
                .andExpect(view().name(StudyController.NEW_STUDY_VIEW))
                .andExpect(model().attributeExists("studySubmitForm"))
                .andExpect(model().hasErrors())
                .andDo(print());
        Optional<Study> byUrl = studys.findByUrl(url);
        assertFalse(byUrl.isPresent());
    }
}